tests/wpt/web-platform-tests/css/css-align-3/self-alignment/place-self-shorthand-005.html

<!DOCTYPE html>
<title>CSS Box Alignment: place-self shorthand - inherit value</title>
<link rel="author" title="Javier Fernandez" href="mailto:jfernandez@igalia.com" />
<link rel="help" href="http://www.w3.org/TR/css3-align/#place-self-property" />
<meta name="assert" content="Check that place-self's 'inherit' value expands to 'align-self' and 'justify-self'." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        place-self: start end;
    }
</style>
<div id="log"></div>
<div id="test"></div>
<script>
    var child = document.createElement("div");
    document.getElementById("test").appendChild(child);
    child.setAttribute("style", "place-self: inherit");
    var style = getComputedStyle(child);

    test(function() {
        assert_equals(style.getPropertyValue("align-self"),
            "start", "place-self resolved value for align-self");
    }, "Check place-self: inherit - align-self resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("justify-self"),
           "end", "place-self resolved value for justify-self");
    }, "Check place-self: inherit - justify-self resolved value");
</script>
